more fixes

+++ src/misc.c
@@ -2141,7 +2141,7 @@ void *rig_get_function_ptr(rig_model_t rig_model,
 
 // negative return indicates error
 // watch out for integer values that may be negative
-int rig_get_caps_int(rig_model_t rig_model, enum rig_caps_int_e rig_caps)
+long rig_get_caps_int(rig_model_t rig_model, enum rig_caps_int_e rig_caps)
 {
     const struct rig_caps *caps = rig_get_caps(rig_model);
